// src/types/ficha-from-input.ts
export type FichaInputData = {
  nome: string;
  email?: string;
  peso: number;
  peito: number;
  abdomen: number;
  ombros: number;
  quadricepsEsquerdo: number;
  quadricepsDireito: number;
  panturrilhaEsquerda: number;
  panturrilhaDireita: number;
  bicepsEsquerdo: number;
  bicepsDireito: number;
};